Refactored AuthService to be more readable and fixed minor bugs.

+++ src/main/java/pl/edu/pw/ee/cookbookserver/service/impl/AuthServiceImpl.java
@@ -36,33 +36,37 @@ public class AuthServiceImpl implements AuthService {
     @Override
     public ResponseEntity register(JSONObject payload, String origin) throws JSONException {
         String emailKey = PayloadKey.EMAIL.value();
-        if (!payload.has(emailKey) || payload.isNull(emailKey)) {
+        if (!payload.has(emailKey) || payload.isNull(emailKey) || payload.getString(emailKey).isEmpty()) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_EMAIL.value());
         }
-
-        String usernameKey = PayloadKey.USERNAME.value();
-        if (!payload.has(usernameKey) || payload.isNull(usernameKey)) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_USERNAME.value());
-        }
-
-        String passwordKey = PayloadKey.PASSWORD.value();
-        if (!payload.has(passwordKey) || payload.isNull(passwordKey)) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_PASSWORD.value());
-        }
-
         String email = payload.getString(emailKey);
+        if (!email.matches("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*$")) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.INVALID_EMAIL.value());
+        }
         User user = userRepository.findByEmail(email).orElse(null);
         if (user != null) {
             return ResponseEntity.status(HttpStatus.CONFLICT).body(ResponseMessage.EMAIL_OCCUPIED.value());
         }
 
+        String usernameKey = PayloadKey.USERNAME.value();
+        if (!payload.has(usernameKey) || payload.isNull(usernameKey) || payload.getString(usernameKey).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_USERNAME.value());
+        }
         String username = payload.getString(usernameKey);
         user = userRepository.findByUsername(username).orElse(null);
         if (user != null) {
             return ResponseEntity.status(HttpStatus.CONFLICT).body(ResponseMessage.USERNAME_OCCUPIED.value());
         }
 
+        String passwordKey = PayloadKey.PASSWORD.value();
+        if (!payload.has(passwordKey) || payload.isNull(passwordKey) || payload.getString(passwordKey).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_PASSWORD.value());
+        }
         String password = payload.getString(passwordKey);
+        if (password.length() < 8) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.PASSWORD_TOO_SHORT.value());
+        }
+
         user = new User();
         user.setUsername(username);
         user.setPassword(new BCryptPasswordEncoder().encode(password));
@@ -73,7 +77,7 @@ public class AuthServiceImpl implements AuthService {
         user.setEnabled(false);
         userRepository.save(user);
 
-        Token token = createAccessToken(user);
+        Token token = createToken(user);
         mailService.sendAccountActivationMessage(origin, token);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
@@ -81,12 +85,12 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     public ResponseEntity verify(JSONObject payload) throws JSONException {
-        String tokenKey = PayloadKey.TOKEN.value();
-        if (!payload.has(tokenKey) || payload.isNull(tokenKey)) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_TOKEN.value());
+        String uuidKey = PayloadKey.UUID.value();
+        if (!payload.has(uuidKey) || payload.isNull(uuidKey) || payload.getString(uuidKey).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_UUID.value());
         }
-
-        Token token = tokenRepository.findByUuid(payload.getString(tokenKey)).orElse(null);
+        String uuid = payload.getString(uuidKey);
+        Token token = tokenRepository.findByUuid(uuid).orElse(null);
         if (token == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
@@ -105,17 +109,16 @@ public class AuthServiceImpl implements AuthService {
     @Override
     public ResponseEntity reset(JSONObject payload, String origin) throws JSONException {
         String usernameKey = PayloadKey.USERNAME.value();
-        if (!payload.has(usernameKey) || payload.isNull(usernameKey)) {
+        if (!payload.has(usernameKey) || payload.isNull(usernameKey) || payload.getString(usernameKey).isEmpty()) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_USERNAME.value());
         }
-
         String username = payload.getString(usernameKey);
         User user = userRepository.findByUsernameOrEmail(username, username).orElse(null);
         if (user == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
 
-        Token token = createAccessToken(user);
+        Token token = createToken(user);
         mailService.sendPasswordResetMessage(origin, token);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
@@ -123,17 +126,12 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     public ResponseEntity confirm(JSONObject payload) throws JSONException {
-        String tokenKey = PayloadKey.TOKEN.value();
-        if (!payload.has(tokenKey) || payload.isNull(tokenKey)) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_TOKEN.value());
+        String uuidKey = PayloadKey.UUID.value();
+        if (!payload.has(uuidKey) || payload.isNull(uuidKey) || payload.getString(uuidKey).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_UUID.value());
         }
-
-        String passwordKey = PayloadKey.PASSWORD.value();
-        if (!payload.has(passwordKey) || payload.isNull(passwordKey)) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_PASSWORD.value());
-        }
-
-        Token token = tokenRepository.findByUuid(payload.getString(tokenKey)).orElse(null);
+        String uuid = payload.getString(uuidKey);
+        Token token = tokenRepository.findByUuid(uuid).orElse(null);
         if (token == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
@@ -141,15 +139,24 @@ public class AuthServiceImpl implements AuthService {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.TOKEN_EXPIRED.value());
         }
 
+        String passwordKey = PayloadKey.PASSWORD.value();
+        if (!payload.has(passwordKey) || payload.isNull(passwordKey) || payload.getString(passwordKey).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.MISSING_PASSWORD.value());
+        }
+        String password = payload.getString(passwordKey);
+        if (password.length() < 8) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseMessage.PASSWORD_TOO_SHORT.value());
+        }
+
         User user = token.getUser();
-        user.setPassword(new BCryptPasswordEncoder().encode(payload.getString(passwordKey)));
+        user.setPassword(new BCryptPasswordEncoder().encode(password));
         userRepository.save(user);
         tokenRepository.delete(token);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
-    private Token createAccessToken(User user) {
+    private Token createToken(User user) {
         Token token = new Token();
         token.setUser(user);
         token.setUuid(UUID.randomUUID().toString());
